// src/lib/api/generated.ts
export interface LatestEvent {
  /** @maxLength 50 */
  event_slug?: string;
  readonly thumbnail?: string;
}

export interface EventPin {
  /** @maxLength 50 */
  event_slug?: string;
  readonly thumbnail?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  topic: string;
  statement?: string;
  start_datetime?: string;
  finish_datetime?: string;
}

export interface EventTicket {
  readonly id?: number;
  /**
     * @minLength 1
     * @maxLength 300
     */
  provider_name: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  event_name: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  event_category: string;
  /**
     * @minLength 1
     * @maxLength 15
     */
  phone_number: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  attachment: string;
  /** @minLength 1 */
  description: string;
}